AOS-6020 Simpler .ao-config.json
AOS-6020 Work in progress, does not build

AOS-6020 bug fix

AOS-6020 fixing/removing tests that made no sense

AOS-6020 Some code review fixes

AOS-6020 improved logging and user feedback when loading ao config file

AOS-6020 removing unneeded checks for cluster existence

AOS-6020 fjerne unødvendig InitClusters-kall
