+++ pkg/config/openshift_test.go
@@ -85,63 +85,3 @@ func TestGetToken(t *testing.T) {
 		ts.Close()
 	}
 }
-
-func TestAOConfig_InitClusters(t *testing.T) {
-
-	aoConfig := basicAOConfig
-	aoConfig.ClusterURLPattern = "%s"
-	aoConfig.BooberURLPattern = "%s"
-	aoConfig.GoboURLPattern = "%s"
-	aoConfig.AvailableClusters = []string{}
-	aoConfig.PreferredAPIClusters = []string{}
-
-	type TestCase struct {
-		Name      string
-		Reachable bool
-	}
-
-	testCases := []TestCase{
-		{"prod", false},
-		{"test", false},
-		{"utv", true},
-		{"qa", true},
-	}
-
-	testMap := make(map[string]TestCase)
-
-	var testServers []*httptest.Server
-	for _, test := range testCases {
-		reachable := test.Reachable
-		ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
-			if reachable {
-				w.WriteHeader(http.StatusOK)
-			} else {
-				w.WriteHeader(http.StatusInternalServerError)
-			}
-		}))
-		testServers = append(testServers, ts)
-		aoConfig.AvailableClusters = append(aoConfig.AvailableClusters, ts.URL)
-		testMap[ts.URL] = test
-	}
-
-	noHost := "http://nohost:8080"
-	aoConfig.AvailableClusters = append(aoConfig.AvailableClusters, noHost)
-	testMap[noHost] = TestCase{
-		Reachable: false,
-		Name:      "no host",
-	}
-
-	// Test
-	aoConfig.InitClusters()
-	for _, c := range aoConfig.Clusters {
-		test := testMap[c.Name]
-		assert.Equal(t, test.Reachable, c.Reachable)
-
-		// Since clusterUrlPattern is %s then name and url should be equal
-		assert.Equal(t, c.Name, c.URL)
-	}
-
-	for _, t := range testServers {
-		t.Close()
-	}
-}
